Add the ability to "track" and modify/delete spots.
By default, a callsign is only expected to occur once per band, and if
they show up at a new frequency within the band, the old spot will be
moved to the new frequency. If this is disabled with
`-one-per-band false`, updating will only happen if there is a new
spot on the same frequency or within 1kHz.

By default, a spot from the cluster with a comment containing "QRT"
will cause the spot to be deleted from the Flex. This can be disabled
with `-qrt false`.